Remove baseitem moving functionality to item
